requirements.txt
requests
supabase 
python-dotenv==1.1.0
